refactor: move available_port function to a different module
